// frontend/src/app/core/models/clan.model.ts
export interface Clan {
  id: number;
  nombre: string;
  tag: string;
  ranking: number;
  riqueza: number;
  cantidadMiembros: number;
  fechaCreacion: string;
  descripcion: string;
  nivel: number;
  victorias: number;
  derrotas: number;
  territoriosControlados: number;
  color: string;
  totalAliados: number;
  guerrasActivas: number;
}